requirements.txt
Flask
numpy
Pillow
typing